requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==1.1.0
reportlab==4.0.7
aiofiles==23.2.1
jinja2==3.1.2
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
lxml==4.9.3
alembic==1.13.1
